Initial data manipulation step
Started off modifying the data and playing around with it.

Removed the sequence field in the dataset as a pandas frame enumerates the rows by default. I added the modified csv to the repo. Use this one from now on!

Created and applied mappings to replace the non-numerical fields with integers.

Now we need to filter through the information that is useful and useless, judging from the relevant graphs.