Final version of Linear Regression Model and Decision Tree/Random Forest models.
Fixed and ran the hyperparameter training for Random Forests and fixed the comments on both models.

Note:
Both the linear regression model and polynomial regression model are found in the Linear_Regression.py file.

Both the decision tree and random forest models are found in the Decision_tree.py file.